Split app and launcher
