+++ examples/performance-subscribe/src/PerformanceTestApp.h
@@ -22,10 +22,6 @@
 #include <string>
 #include <vector>
 
-namespace velocitas {
-class Vehicle;
-} // namespace velocitas
-
 namespace example {
 
 /**
